# docs/swagger.yaml
definitions:
  dtos.HealthResponse:
    properties:
      serviceName:
        type: string
      status:
        type: string
      timestamp:
        type: string
    type: object
  dtos.UpdateProfileRequest:
    properties:
      displayName:
        type: string
      pictureUrl:
        type: string
    required:
    - displayName
    type: object
  dtos.UpdateRoleRequest:
    properties:
      role:
        type: string
    required:
    - role
    type: object
  dtos.UserResponse:
    properties:
      displayName:
        type: string
      email:
        type: string
      id:
        type: string
      pictureUrl:
        type: string
      role:
        type: string
    type: object
info:
  contact: {}
  description: REST API for the Subglutee user service
  title: User Service API
  version: "1.0"
paths:
  /api/v1/users:
    get:
      parameters:
      - description: limit (default 10)
        in: query
        name: limit
        type: integer
      - description: offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.UserResponse'
            type: array
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      summary: list all users
      tags:
      - admin
  /api/v1/users/{id}:
    delete:
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      summary: delete a user
      tags:
      - admin
    get:
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.UserResponse'
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      summary: get a user by id
      tags:
      - admin
    patch:
      consumes:
      - application/json
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      - description: new role
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.UserResponse'
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      summary: change a user's role
      tags:
      - admin
  /api/v1/users/me:
    delete:
      responses:
        "204":
          description: No Content
      summary: delete my own account
      tags:
      - users
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.UserResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: get my profile
      tags:
      - users
    patch:
      consumes:
      - application/json
      parameters:
      - description: profile
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.UserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: update my profile
      tags:
      - users
  /health:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.HealthResponse'
      summary: health check
      tags:
      - ops
swagger: "2.0"